curio-django/requirements.txt
Django==1.11.4
django-rest-framework==0.1.0
django_extensions==1.8.1
django-cors-headers==2.1.0
django-filter==1.0.4
